feat: v1.0.0 — stable public API, promote clock package, pre-release polish
API surface audit (46 confirmed fixes across 23 packages) ahead of the v1.0.0
stability commitment. Key changes:

- Promote internal/clock to a top-level public `clock` package so callers can
  inject custom clocks and write deterministic tests; internal/clock becomes a
  backward-compatible type-alias shim.
- CircuitError.Err field made private (err); add NewCircuitError constructor so
  the struct remains constructible externally.
- ratelimit/composite: extract options.go (WithClock, WithRecorder,
  WithOnDecision) to match the option-file convention in every other package.
- retry/backoff: mutex-guard shared rand.Rand in jitter and decorrelated
  strategies; fix decorrelated d.prev race (H-13/H-14 from adversarial audit).
- Uniform godoc on every exported identifier that was missing a comment.
- Constructor validation (zero/negative capacity, window, rate) across
  slidingwindow, fixedwindow, gcra, tokenbucket.
- examples/tiered: fix New() call to pass []Tier slice not variadic Tier values.
- Update CHANGELOG [Unreleased] → [1.0.0] 2026-07-21; README stability section;
  docs/STABILITY.md: "pre-1.0 (v0.x)" → "stable (v1.x)".

+++ bulkhead/bulkhead_test.go
@@ -40,12 +40,12 @@ func TestBulkhead_AllowUpToConcurrencyLimit(t *testing.T) {
 	// Wait until all limit slots are occupied.
 	deadline := time.Now().Add(2 * time.Second)
 	for time.Now().Before(deadline) {
-		if b.Inflight() == int64(limit) {
+		if b.Inflight() == limit {
 			break
 		}
 		time.Sleep(5 * time.Millisecond)
 	}
-	if got := b.Inflight(); got != int64(limit) {
+	if got := b.Inflight(); got != limit {
 		t.Fatalf("expected %d inflight, got %d", limit, got)
 	}
 
@@ -234,7 +234,7 @@ func TestBulkhead_MetricsAccurate(t *testing.T) {
 	}
 	started.Wait()
 
-	if got := b.Inflight(); got != int64(limit) {
+	if got := b.Inflight(); got != limit {
 		t.Fatalf("expected %d inflight, got %d", limit, got)
 	}
 
@@ -253,11 +253,11 @@ func TestBulkhead_MetricsAccurate(t *testing.T) {
 	}
 	wg.Wait()
 
-	expectedRejected := int64(total - limit)
+	expectedRejected := total - limit
 	if got := b.Rejected(); got != expectedRejected {
 		t.Fatalf("expected %d rejected, got %d", expectedRejected, got)
 	}
-	if got := rejected.Load(); got != expectedRejected {
+	if got := int(rejected.Load()); got != expectedRejected {
 		t.Fatalf("locally counted %d rejections, expected %d", got, expectedRejected)
 	}
